backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

SQLAlchemy[asyncio]==2.0.32
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2

python-jose[cryptography]==3.3.0

# ใช้ passlib 1.7.4 + bcrypt 3.2.2 เพื่อไม่ให้ขึ้น "(trapped) error reading bcrypt version"
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

redis==5.0.7

